refactor: decompose oversized files (>300 LOC) (#442)
Split large Go source files into smaller, focused modules:

internal/ui/tui.go (1095 → 413 LOC):
- tui_view.go: View rendering (187 LOC)
- tui_commands.go: Tea Cmd functions (148 LOC)
- tui_filter.go: Filtering, sorting, utilities (158 LOC)
- tui_keys.go: Keyboard handler (194 LOC)

internal/vault/entry.go (1033 → 209 LOC):
- entry_validate.go: Path validation, storage paths (156 LOC)
- entry_readwrite.go: CRUD operations, config, cloning (541 LOC)
- entry_canary.go: Canary/honeytoken operations (113 LOC)

internal/git/git.go (993 → 258 LOC):
- git_push.go: Push operations, SSH auth (145 LOC)
- git_pull.go: Pull, sync, conflict resolution (232 LOC)
- git_staging.go: Staging, index operations (148 LOC)
- git_remote.go: Remote management (74 LOC)
- git_util.go: Log, utilities (83 LOC)

All existing tests pass. No logic or API changes.
